made TaskListApp in it's own page
