+++ src/main.dart
@@ -1,5 +1,5 @@
 import 'dart:io';
-
+import 'TaskListApp.dart';
 void main(List<String> args) {
   List<TaskListApp> tasks = [];
   print("♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡");
@@ -118,40 +118,3 @@ void main(List<String> args) {
   } while (answer != 5);
 }
 
-class TaskListApp {
-  int id;
-  String taskname;
-  String description;
-  bool isCompleted;
-
-  // Default constructor
-  TaskListApp()
-      : this.withDetails(
-            id: 0, taskname: '', description: '', isCompleted: false);
-
-  // Constructor with named parameters
-  TaskListApp.withDetails({
-    required this.id,
-    required this.taskname,
-    required this.description,
-    required this.isCompleted,
-  });
-
-  @override
-  String toString() {
-    String finished = isCompleted ? 'Yes' : 'No';
-    print("♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡♡");
-    return 'Task number $id \ntaskname: $taskname \ndescription: $description \nFinished: $finished';
-
-  }
-}
-
-extension IterableExtensions<T> on Iterable<T> {
-  T? firstWhereOrNull(bool Function(T) test) {
-    try {
-      return firstWhere(test);
-    } catch (e) {
-      return null;
-    }
-  }
-}
